+++ app/src/main/java/woowacourse/shopping/presentation/shopping/BindingAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.presentation.product
+package woowacourse.shopping.presentation.shopping
 
 import android.widget.ImageView
 import androidx.databinding.BindingAdapter
