+++ app/src/main/java/woowacourse/shopping/presentation/shopping/product/ProductUi.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.presentation.product
+package woowacourse.shopping.presentation.shopping.product
 
 data class ProductUi(
     val id: Long,
