+++ app/src/main/res/layout/activity_product.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.product.ProductActivity">
+        tools:context=".presentation.shopping.product.ProductActivity">
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/fragment_container_shopping"
