+++ app/src/main/res/layout/fragment_product_detail.xml
@@ -5,14 +5,14 @@
 
         <variable
             name="vm"
-            type="woowacourse.shopping.presentation.product.ProductDetailViewModel" />
+            type="woowacourse.shopping.presentation.shopping.detail.ProductDetailViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.product.ProductDetailFragment">
+        tools:context=".presentation.shopping.detail.ProductDetailFragment">
 
         <ImageView
             android:id="@+id/iv_product"
